Project beginning
Uploading initial files for the project which includes the to be developed House class as an example and overall project structure.
Current Makefile is only a placeholder and is unwieldly in regards to the growth of this project. Will be enhanced further in the future with the use of wildcards and sub-makefiles, as well as rules definitions.