Implemented rudimentary rocket
Rocket has 3 states, ARMED where it is waiting a command to be shot, FLYING where it is flying to the destination, and finally, EXPLODING, which must be intuitive enough. Just click to fire away.

No collision in this stage yet, since the location of other objects and their behavior hasn't been defined yet.